TRAIN-DASHBOARD-V2/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
supabase==1.0.3
pandas==2.1.4
pyecharts==2.0.4
httpx==0.23.3
gotrue==1.0.4 